feat: add jwt config / add login, signup api (not complete) / change package name 'auctionnara' to boltauction
