--- src/main/java/com/neoga/autionnara/AutionnaraApplication.java
@@ -1,13 +1,0 @@
-package com.neoga.autionnara;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AutionnaraApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AutionnaraApplication.class, args);
-    }
-}
-
