--- src/main/java/com/neoga/autionnara/category/domain/Category.java
@@ -1,17 +1,0 @@
-package com.neoga.autionnara.category.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@Getter
-@Setter
-@Entity
-public class Category {
-    @Column(name="category_id")
-    @Id
-    @GeneratedValue(strategy= GenerationType.AUTO)
-    private Long id;
-    private String name;
-}
