--- src/main/java/com/neoga/autionnara/chat/domain/Room.java
@@ -1,18 +1,0 @@
-package com.neoga.autionnara.chat.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@Entity
-public class Room {
-    @Column(name="chat_id")
-    @Id
-    @GeneratedValue(strategy= GenerationType.AUTO)
-    private Long id;
-    private LocalDateTime create_date;
-}
