--- src/main/java/com/neoga/autionnara/chat/domain/RoomJoin.java
@@ -1,25 +1,0 @@
-package com.neoga.autionnara.chat.domain;
-
-
-import com.neoga.autionnara.memberstore.member.domain.Members;
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@Getter
-@Setter
-@Entity
-@Table(name="room_join")
-public class RoomJoin {
-    @Column(name="room_join_id")
-    @Id
-    @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long id;
-    @ManyToOne
-    @JoinColumn(name = "room_id")
-    private Room room;
-    @ManyToOne
-    @JoinColumn(name = "member_id")
-    private Members members;
-}
