--- src/main/java/com/neoga/autionnara/memberstore/member/domain/Role.java
@@ -1,5 +1,0 @@
-package com.neoga.autionnara.memberstore.member.domain;
-
-public enum Role {
-    ROLE_MEMBER, ROLE_ADMIN
-}
