--- src/main/java/com/neoga/autionnara/order/domain/Orders.java
@@ -1,24 +1,0 @@
-package com.neoga.autionnara.order.domain;
-
-import com.neoga.autionnara.item.domain.Item;
-import com.neoga.autionnara.memberstore.member.domain.Members;
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@Getter
-@Setter
-@Entity
-public class Orders {
-    @Column(name="order_id")
-    @Id
-    @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long id;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Members members;
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "item_id")
-    private Item item;
-}
