+++ src/main/java/com/neoga/boltacution/AutionnaraApplication.java
@@ -1,0 +1,13 @@
+package com.neoga.boltacution;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AutionnaraApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AutionnaraApplication.class, args);
+    }
+}
+
