+++ src/main/java/com/neoga/boltacution/bid/domain/Bid.java
@@ -1,0 +1,24 @@
+package com.neoga.boltacution.bid.domain;
+
+import com.neoga.boltacution.item.domain.Item;
+import com.neoga.boltacution.memberstore.member.domain.Members;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Getter
+@Setter
+@Entity
+public class Bid {
+    @Column(name="bid_id")
+    @Id
+    @GeneratedValue(strategy=GenerationType.AUTO)
+    private Long id;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Members members;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "item_id")
+    private Item item;
+}
