+++ src/main/java/com/neoga/boltacution/chat/domain/Room.java
@@ -1,0 +1,18 @@
+package com.neoga.boltacution.chat.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@Entity
+public class Room {
+    @Column(name="chat_id")
+    @Id
+    @GeneratedValue(strategy= GenerationType.AUTO)
+    private Long id;
+    private LocalDateTime create_date;
+}
