+++ src/main/java/com/neoga/boltacution/exception/CEmailLoginFailedException.java
@@ -1,0 +1,15 @@
+package com.neoga.boltacution.exception;
+
+public class CEmailLoginFailedException extends RuntimeException{
+    public CEmailLoginFailedException(String msg, Throwable t) {
+        super(msg, t);
+    }
+
+    public CEmailLoginFailedException(String msg) {
+        super(msg);
+    }
+
+    public CEmailLoginFailedException() {
+        super();
+    }
+}
