+++ src/main/java/com/neoga/boltacution/memberstore/member/domain/Role.java
@@ -1,0 +1,5 @@
+package com.neoga.boltacution.memberstore.member.domain;
+
+public enum Role {
+    USER, ADMIN
+}
