+++ src/main/java/com/neoga/boltacution/memberstore/review/domain/Review.java
@@ -1,0 +1,25 @@
+package com.neoga.boltacution.memberstore.review.domain;
+
+import com.neoga.boltacution.memberstore.member.domain.Members;
+import com.neoga.boltacution.memberstore.store.domain.Store;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Getter
+@Setter
+@Entity
+public class Review {
+    @Column(name="review_id")
+    @Id
+    @GeneratedValue(strategy=GenerationType.AUTO)
+    private Long id;
+    private String content;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "register_id")
+    private Members register;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "store_id")
+    private Store store;
+}
