+++ src/main/java/com/neoga/boltacution/memberstore/store/domain/Store.java
@@ -1,0 +1,19 @@
+package com.neoga.boltacution.memberstore.store.domain;
+
+import com.neoga.boltacution.memberstore.member.domain.Members;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Getter
+@Setter
+@Entity
+public class Store {
+    @Column(name="store_id")
+    @Id
+    @GeneratedValue(strategy=GenerationType.AUTO)
+    private Long id;
+    @OneToOne(mappedBy="store", fetch = FetchType.LAZY)
+    private Members members;
+}
