+++ src/main/java/com/neoga/boltacution/security/domain/SecurityMember.java
@@ -1,0 +1,56 @@
+package com.neoga.boltacution.security.domain;
+
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.neoga.boltacution.memberstore.member.domain.Members;
+import com.neoga.boltacution.memberstore.member.domain.Role;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class SecurityMember extends User {
+    private Members member;
+
+    public SecurityMember(Members member) {
+        super(member.getEmail(), member.getPasswd(), authorities(member.getRole()));
+        this.member = member;
+    }
+
+    public Members getMember() {
+        return member;
+    }
+
+    private static Collection<? extends GrantedAuthority> authorities(List<Role> roles) {
+        return roles.stream()
+                .map(r -> new SimpleGrantedAuthority("ROLE_" + r.name()))
+                .collect(Collectors.toSet());
+    }
+
+    @Override
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    public boolean isEnabled() {
+        return true;
+    }
+}
