--- src/test/java/com/neoga/autionnara/AutionnaraApplicationTests.java
@@ -1,13 +1,0 @@
-package com.neoga.autionnara;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AutionnaraApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
