+++ src/test/java/com/neoga/boltacution/AutionnaraApplicationTests.java
@@ -1,0 +1,13 @@
+package com.neoga.boltacution;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AutionnaraApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
